# backend/api.yml
openapi: 3.0.3
info:
  title: Aspirant AI API
  version: 1.0.0
  description: API documentation for the Aspirant AI backend
paths:
...
  /api/community/contributors/profile/{username}/:
    get:
      operationId: community_contributors_profile_retrieve
      description: GET /api/community/contributors/profile/<username>/
      parameters:
      - in: path
        name: username
        schema:
          type: string
        required: true
      tags:
      - community
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          description: No response body
...
components:
...
  securitySchemes:
    jwtAuth:
      type: apiKey
      in: header
      name: Authorization
      description: Token-based authentication with required prefix "Token"